src/cli/pyproject.toml
[project]
name = "qtcloud-asset-cli"
version = "0.1.0"
description = "量潮数字资产云 CLI"
requires-python = ">=3.12"
dependencies = [
    "typer",
    "pyyaml",
    "pydantic",
    "rich",
    "requests",
    "python-dotenv",
]

[project.optional-dependencies]
dev = [
    "pytest",
    "pytest-cov",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/cli"]

[tool.pytest.ini_options]
pythonpath = ["."]
